tests for new utils
